+++ src/components/auth/LoginButton.tsx
@@ -3,26 +3,9 @@
 import KakaoSymbol from './KakaoSymbol';
 import GoogleLogo from './GoogleLogo';
 import styles from './LoginButton.module.css';
-import { createClient } from '@/utils/supabase/client';
-
-const getURL = () => {
-  let url = process?.env?.NEXT_PUBLIC_SITE_URL ?? process?.env?.NEXT_PUBLIC_VERCEL_URL ?? 'http://localhost:3000/';
-  url = url.includes('http') ? url : `https://${url}`;
-  url = url.charAt(url.length - 1) === '/' ? url : `${url}/`;
-  return `${url}/auth/callback`;
-};
+import { logInWithOAuth } from '@/api/auth';
 
 const LoginButton = () => {
-  const logInWithOAuth = async (provider: 'google' | 'kakao') => {
-    const supabase = createClient();
-    await supabase.auth.signInWithOAuth({
-      provider,
-      options: {
-        redirectTo: getURL()
-      }
-    });
-  };
-
   return (
     <form className={styles.form}>
       <button className={styles.kakao} onClick={() => logInWithOAuth('kakao')} role="button">
