+++ src/components/header/Header.tsx
@@ -1,9 +1,9 @@
 import { createClient } from '@/utils/supabase/server';
 import Logout from '../auth/LogoutButton';
-import UserProfile from '../auth/UserProfileButton';
 import styles from './Header.module.css';
 import Link from 'next/link';
-import { MeetingModalButton } from '../auth/MeetingModalButton';
+import { MeetingModalButton } from './MeetingModalButton';
+import UserProfile from './UserProfileButton';
 
 const Header = async () => {
   const supabase = createClient();
@@ -14,11 +14,11 @@ const Header = async () => {
 
   return (
     <header className={styles.header}>
-      {user ? <MeetingModalButton /> : null}
-      <Link href="/">owl-link</Link>
+      <Link href="http://localhost:3000/">owl-link</Link>
       <div className={styles.menu}>
         {user ? (
           <>
+            <MeetingModalButton />
             <UserProfile />
             <Logout />
           </>
